Returning units
Units no longer crash the game with the defines fixed. However, only returning the 1980 and generic templates - the old ones still do, and it's not like they're needed. Also, only activating OOB for Afghanistan, so the game runs faster, if you need OOB activated elsewhere, remove the # in the country's history file before the OOB line.